requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
magic-filter==1.0.9
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
psycopg2-binary==2.9.6
pydantic==1.10.11
python-dotenv==1.0.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
yarl==1.9.2
